requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psutil==6.0.0
browser-history==0.4.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pywinctl
pyautogui
pygetwindow
selenium
webdriver_manager
python-xlib
pathlib
deptry==0.17.0